influencer page

+++ app/utils.ts
@@ -28,7 +28,7 @@ export class FilterArrayPipe implements PipeTransform {
 @Pipe({ name: 'fmtNum' })
 export class FormatNumberPipe implements PipeTransform {
 	transform(n, fn): string {
-		n = parseInt(n);
+		n = parseFloat(n);
 		if (n === 0 || isNaN(n)) return 'N/A';
 		if (n >= 1e6) {
 			return trimNumber(n / 1e6) + 'M';
@@ -40,7 +40,7 @@ export class FormatNumberPipe implements PipeTransform {
 	}
 }
 
-function trimNumber(n: number): string { return n.toFixed(2).replace(/\.0\d?$/, '');}
+function trimNumber(n: number): string { return n.toFixed(2).replace(/\.0\d?$/, ''); }
 
 export function FilterByProps(kw: string | null, it: Object, ...props: string[]): boolean {
 	if (!kw || !it) return true;
